Correct MajorOffset_Size_DLA computation
Wrong cast. Number of bytes of a major Offset in DynLock_Area can exceed
255 bytes (max 2pow15)

+++ src/nfc/tags/rw_t2t_ndef.cc
@@ -677,7 +677,8 @@ static void rw_t2t_handle_tlv_detect_rsp(uint8_t* p_data) {
                   p_t2t->mem_tlv[p_t2t->num_mem_tlvs].offset =
                       (p_t2t->tlv_value[0] >> 4) & 0x0F;
                   p_t2t->mem_tlv[p_t2t->num_mem_tlvs].offset *=
-                      (uint8_t)tags_pow(2, p_t2t->tlv_value[2] & 0x0F);
+                      (uint16_t)tags_pow(2, p_t2t->tlv_value[2] & 0x0F);
+
                   p_t2t->mem_tlv[p_t2t->num_mem_tlvs].offset +=
                       p_t2t->tlv_value[0] & 0x0F;
                   count = p_t2t->tlv_value[1];
